// extension/src/api/types/event.ts
export enum LogEvent {
  USER_ACCEPT = "USER_ACCEPT",
  USER_REJECT = "USER_REJECT",

  USER_LINE_ACCEPT = "USER_LINE_ACCEPT",
  USER_LINE_REJECT = "USER_LINE_REJECT",

  MODEL_GENERATE = "MODEL_GENERATE",
  MODEL_ERROR = "MODEL_ERROR",

  USER_SIGNUP = "USER_SIGNUP",
  USER_LOGIN = "USER_LOGIN",
  USER_LOGOUT = "USER_LOGOUT",

  USER_AUTH_GITHUB = "USER_AUTH_GITHUB",

  USER_LOCKED = "USER_LOCKED",
  USER_UNLOCKED = "USER_UNLOCKED",

  USER_ANSWER_CORRECT = "USER_ANSWER_CORRECT",
  USER_ANSWER_INCORRECT = "USER_ANSWER_INCORRECT",
}

export interface LogData {
  event: LogEvent;
  timeLapse: number;
  metadata: Record<string, any>;
}